% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oswestry_eng}
\alias{oswestry_eng}
\title{Possible answer for each item on the English Oswestry Disability Index.}
\format{
A data frame with 12 rows and 25 variables:
\describe{
  \item{id}{identifyers for each participants}
  \item{date}{date at which the survey was filled in}
  \item{painIntensity}{`Pain Intensity` item in the Oswestry disability index}
  \item{personalCare}{`Personal Care` item in the Oswestry disability index}
  \item{lifting}{`Lifting` item in the Oswestry disability index}
  \item{walking}{`Walking` item in the Oswestry disability index}
  \item{sitting}{`Sitting` item in the Oswestry disability index}
  \item{standing}{`Standing` item in the Oswestry disability index}
  \item{sleeping}{`Sleeping` item in the Oswestry disability index}
  \item{sexLife}{`Sex life` item in the Oswestry disability index}
  \item{socialLife}{`Social life` item in the Oswestry disability index}
  \item{travelling}{`Travelling` item in the Oswestry disability index}
  \item{scored.}{Computed scores for each individual items}
  \item{numerator}{Sum of the individual item scores for each individual}
  \item{denominator}{Maximum attainable score for each individual}
  \item{oswestry}{Oswestry disability index for each individual}
}
}
\usage{
oswestry_eng
}
\description{
Possible answer for each item on the English Oswestry Disability Index.
}
\keyword{datasets}
