% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdi.R
\name{score.bdi}
\alias{score.bdi}
\title{Beck Depression Inventory}
\usage{
score.bdi(
  data,
  obs,
  cols,
  date = NULL,
  keepResponses = FALSE,
  keepScoring = FALSE
)
}
\arguments{
\item{data}{A dataframe containing the survey responses.}

\item{obs}{A string specifying the column containing participants' id.}

\item{cols}{An interger vector specifying the range of the items to score.
Must be of length 21. Note that `unique()` is applied on the vector.
A warning is given if too many NA values are detected in the output.
Additionnaly, errors given in ``rename_impl()` also hint towards cols set incorrectly.}

\item{date}{A string specifying the column containing the date.}

\item{keepResponses}{A logical: should original item responses be kept?}

\item{keepScoring}{A logical: should item scoring be kept?}
}
\value{
A dataframe or tibble containing the scored Beck Depression Inventory.
}
\description{
Returns a tibble containing the decoded scores of the Beck Depression Inventory.
It is possible to keep the original item repsonses and the item scoring if desired.
}
