% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oswestry.R
\name{score.oswestry}
\alias{score.oswestry}
\title{score.oswestry}
\usage{
score.oswestry(
  data,
  obs,
  cols,
  language = "english",
  date = NULL,
  keepResponses = FALSE,
  keepScoring = FALSE
)
}
\arguments{
\item{data}{A dataframe.}

\item{obs}{A string specifying the column containing participants' id.}

\item{cols}{An interger vector specifying the range of the items to score.}

\item{language}{A string specifying the language. Can be either "english" or "french".}

\item{date}{A string specifying the column contianing the date.}

\item{keepResponses}{A logical: should original item responses be kept?}

\item{keepScoring}{A logical: should item scoring be kept?}
}
\value{
A tibble.
}
\description{
score.oswestry
}
\examples{
score.oswestry(
data = dat,
obs = 'Token',
cols = 3:12,
language = 'french')
}
