% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs.R
\name{score.pcs}
\alias{score.pcs}
\title{Pain Catastrophizing Scale}
\usage{
score.pcs(
  data,
  obs,
  cols,
  language = "english",
  date = NULL,
  keepResponses = FALSE,
  keepScoring = FALSE
)
}
\arguments{
\item{data}{A dataframe containing the survey responses.}

\item{obs}{A string specifying the column containing participants' id.}

\item{cols}{An interger vector specifying the range of the items to score.
Must be of length 10. Note that `unique()` is applied on the vector.
A warning is given if too many NA values are detected in the output.
Additionnaly, errors given in ``rename_impl()` also hint towards cols set incorrectly.}

\item{language}{A string specifying the language. Can be either "english" or "french".}

\item{date}{A string specifying the column containing the date.}

\item{keepResponses}{A logical: should original item responses be kept?}

\item{keepScoring}{A logical: should item scoring be kept?}
}
\value{
A dataframe or tibble containing the scored Pain Catastrophizing Scale
}
\description{
Returns a tibble containing the decoded scores of the Pain Catastrophizing Scale.
It is possible to keep the original item repsonses and the item scoring if desired.
}
\references{
Wheeler, C., Williams, A., & Morley, S. J. (2019). Meta-analysis of the psychometric properties of the Pain Catastrophizing Scale and associations with participant characteristics. Pain, 160(9), 1946–1953. https://doi.org/10.1097/j.pain.0000000000001494

       French, D.J., Noël, M., Vigneau, F., French, J.A., Cyr, C., & Evans, R.T. (2005). L'Échelle de dramatisation face à la douleur PCS-CF Adaptation canadienne en langue française de l'échelle « Pain Catastrophizing Scale ». Canadian Journal of Behavioural Science, 37, 181-192.
}
