% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oswestry.R
\name{score.oswestry}
\alias{score.oswestry}
\title{Oswestry Disability Index}
\usage{
score.oswestry(
  data,
  obs,
  cols,
  language = "english",
  date = NULL,
  keepResponses = FALSE,
  keepScoring = FALSE
)
}
\arguments{
\item{data}{A dataframe containing the survey responses.}

\item{obs}{A string specifying the column containing participants' id.}

\item{cols}{An interger vector specifying the range of the items to score.
Must be of length 10. Note that `unique()` is applied on the vector.}

\item{language}{A string specifying the language. Can be either "english" or "french".}

\item{date}{A string specifying the column containing the date.}

\item{keepResponses}{A logical: should original item responses be kept?}

\item{keepScoring}{A logical: should item scoring be kept?}
}
\value{
A dataframe or tibble containing the scored Oswestry Disability Index.
}
\description{
`score_oswestry()` returns a tibble containing the decoded scores of the
Oswestry Disability Index. It is possible to keep the original item repsonses
and the item scoring if desired.
}
\examples{
library(dplyr)
score.oswestry(
data = oswestry_eng,
obs = 'id',
cols = 3:12
)
}
