% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oswestry.R
\name{score.oswestry}
\alias{score.oswestry}
\title{Oswestry Disability Index}
\usage{
score.oswestry(
  data,
  obs,
  cols,
  language = "english",
  date = NULL,
  keepResponses = FALSE,
  keepScoring = FALSE
)
}
\arguments{
\item{data}{A dataframe containing the survey responses.}

\item{obs}{A string specifying the column containing participants' id.}

\item{cols}{An interger vector specifying the range of the items to score.
Must be of length 10. Note that `unique()` is applied on the vector.
A warning is given if too many NA values are detected in the output.
Additionnaly, errors given in ``rename_impl()` also hint towards cols set incorrectly.}

\item{language}{A string specifying the language. Can be either "english" or "french".}

\item{date}{A string specifying the column containing the date.}

\item{keepResponses}{A logical: should original item responses be kept?}

\item{keepScoring}{A logical: should item scoring be kept?}
}
\value{
A dataframe or tibble containing the scored Oswestry Disability Index.
}
\description{
Returns a tibble containing the decoded scores of the Oswestry Disability Index.
It is possible to keep the original item repsonses and the item scoring if desired.
}
\references{
Denis, I., & Fortin, L. (2012). Development of a French-Canadian version of the Oswestry Disability Index: cross-cultural adaptation and validation. Spine, 37(7), E439–E444. https://doi.org/10.1097/BRS.0b013e318233eaf9

       Fairbank J. C. (2014). Why are there different versions of the Oswestry Disability Index?. Journal of neurosurgery. Spine, 20(1), 83–86. https://doi.org/10.3171/2013.9.SPINE13344
}
